Adds menu nav item "Co-Lab Repos".
It may be confusing to the normal user that the Explore page lists all Co-Lab repos with forks and other repos. Although is also interesting having those, this change proposes having two links, "Co-Labs Repos" for the "official" Co-Lab repos. And aside the existing "Explore" (maybe changing name).
I've not tested the change, just orientative. There should be added "{{if .PageIsExplore}} active{{end}}" with the right conditional for this item, and maybe not needed the translation string for "Co-Lab Repos".
(Is needed to re-build some files from templates?)